+++ sinto/arguments.py
@@ -1,10 +1,7 @@
-#! /usr/bin/env python
-
 import sys
 from argparse import ArgumentParser
 import pkg_resources
 from sinto import cli
-import sys
 
 
 version = pkg_resources.require("sinto")[0].version
@@ -59,6 +56,21 @@ parser_filterbarcodes.add_argument(
     required=False,
     type=str,
     default="CB",
+)
+parser_filterbarcodes.add_argument(
+    "--outdir",
+    help='Output file directory',
+    required=False,
+    type=str,
+    default='.',
+)
+parser_filterbarcodes.add_argument(
+    "-s",
+    "--sam",
+    help="Output sam format (default bam output)",
+    required=False,
+    action="store_true",
+    default=False,
 )
 parser_filterbarcodes.set_defaults(func=cli.run_filterbarcodes)
 
@@ -157,6 +169,56 @@ parser_tagtorg.add_argument(
 
 parser_tagtorg.set_defaults(func=cli.run_tagtorg)
 
+# tagtotag
+parser_tagtotag = subparsers.add_parser(
+    "tagtotag",
+    description="Copies BAM entries to a new file while copying a read tag"
+                " to another read tag"
+                " and optionally deleting the originating tag.",
+)
+parser_tagtotag.add_argument(
+    "-b",
+    "--bam",
+    help="Input SAM/BAM file, '-' reads from stdin",
+    required=True,
+    type=str,
+)
+parser_tagtotag.add_argument(
+    "--from",
+    help="Read tag to copy from.",
+    required=True,
+    dest='from_',
+    type=str,
+)
+parser_tagtotag.add_argument(
+    "--to",
+    help="Read tag to copy to.",
+    required=True,
+    type=str,
+)
+parser_tagtotag.add_argument(
+    "--delete",
+    help="Delete originating tag after copy (i.e. move).",
+    action='store_true',
+)
+parser_tagtotag.add_argument(
+    "-o",
+    "--output",
+    help="Output SAM/BAM file, '-' outputs to stdout (default '-')",
+    default="-",
+    type=str,
+)
+parser_tagtotag.add_argument(
+    "-O",
+    "--outputformat",
+    help="Output format. One of 't' (SAM), 'b' (BAM),"
+    " or 'u' (uncompressed BAM) ('t' default)",
+    default="t",
+)
+
+
+parser_tagtotag.set_defaults(func=cli.run_tagtotag)
+
 # fragments
 parser_fragments = subparsers.add_parser(
     "fragments", description="Create ATAC-seq fragment file from BAM file"
@@ -224,7 +286,7 @@ parser_fragments.add_argument(
     "--use_chrom",
     help="""
     Regular expression used to match chromosomes to be included in output.
-    Default is "(?i)^chr" to match all chromosomes starting with "chr", 
+    Default is "(?i)^chr" to match all chromosomes starting with "chr",
     case insensitive
     """,
     required=False,
@@ -235,11 +297,22 @@ parser_fragments.add_argument(
     "--max_distance",
     help="""
     Maximum distance between integration sites for the fragment to be retained.
-    Allows filtering of implausible fragments that likely result from incorrect 
+    Allows filtering of implausible fragments that likely result from incorrect
     mapping positions. Default is 5000 bp.
     """,
     required=False,
     default=5000,
+    type=int,
+)
+parser_fragments.add_argument(
+    "--min_distance",
+    help="""
+    Minimum distance between integration sites for the fragment to be retained.
+    Allows filtering of implausible fragments that likely result from incorrect
+    mapping positions. Default is 10 bp.
+    """,
+    required=False,
+    default=10,
     type=int,
 )
 parser_fragments.add_argument(
@@ -251,6 +324,34 @@ parser_fragments.add_argument(
     required=False,
     default=500000,
     type=int,
+)
+parser_fragments.add_argument(
+    "--shift_plus",
+    help="""
+    Number of bases to shift Tn5 insertion position by on the forward strand.
+    """,
+    required=False,
+    default=4,
+    type=int,
+)
+parser_fragments.add_argument(
+    "--shift_minus",
+    help="""
+    Number of bases to shift Tn5 insertion position by on the reverse strand.
+    """,
+    required=False,
+    default=-5,
+    type=int,
+)
+parser_fragments.add_argument(
+    "--collapse_within",
+    help="""
+    Take cell barcode into account when collapsing duplicate fragments. Setting this
+    flag means that fragments with the same coordinates can be identified provided
+    they originate from a different cell barcode.
+    """,
+    action="store_true",
+    default=False,
 )
 parser_fragments.set_defaults(func=cli.run_fragments)
 
@@ -270,7 +371,6 @@ parser_blocks.add_argument(
     required=True,
     type=str,
 )
-
 parser_blocks.add_argument(
     "-m",
     "--min_mapq",
@@ -327,10 +427,121 @@ parser_blocks.add_argument(
 )
 parser_blocks.set_defaults(func=cli.run_blocks)
 
-if len(sys.argv[1:]) == 0:
-    parser.print_help()
-    parser.exit()
-else:
-    options = parser.parse_args()
-    sys.setrecursionlimit(10000)
-    options.func(options)
+# barcodes
+parser_barcode = subparsers.add_parser(
+    "barcode", description="Add cell barcode sequences to read names in FASTQ file."
+)
+parser_barcode.add_argument(
+    "--barcode_fastq", help="FASTQ file containing cell barcode sequences", required=True, type=str
+)
+parser_barcode.add_argument(
+    "--read1", help="FASTQ file containing read 1", required=True, type=str
+)
+parser_barcode.add_argument(
+    "--read2", help="FASTQ file containing read 2", required=False, type=str, default=None
+)
+parser_barcode.add_argument(
+    "-b", "--bases", help="Number of bases to extract from barcode-containing FASTQ", required=True, type=int
+)
+parser_barcode.add_argument(
+    "--prefix", help="Prefix to add to cell barcodes", required=False, type=str, default=""
+)
+parser_barcode.add_argument(
+    "--suffix", help="Suffix to add to cell barcodes", required=False, type=str, default=""
+)
+parser_barcode.set_defaults(func=cli.run_barcode)
+
+# tagtoname
+parser_tagtoname = subparsers.add_parser(
+    "tagtoname",
+    description="""
+    Copy cell barcode sequences from tag to read names. Cell barcodes will be added
+    as a readname prefix, followed by ":"
+    """
+)
+parser_tagtoname.add_argument(
+    "-b",
+    "--bam",
+    help="Input SAM/BAM file, '-' reads from stdin",
+    required=True,
+    type=str,
+)
+parser_tagtoname.add_argument(
+    "-o",
+    "--output",
+    help="Output SAM/BAM file, '-' outputs to stdout (default '-')",
+    default="-",
+    type=str,
+)
+parser_tagtoname.add_argument(
+    "-O",
+    "--outputformat",
+    help="Output format. One of 't' (SAM), 'b' (BAM),"
+    " or 'u' (uncompressed BAM) ('t' default)",
+    default="t",
+)
+parser_tagtoname.add_argument(
+    "--tag",
+    help="Read tag to copy from.",
+    required=False,
+    default="CB",
+    type=str,
+)
+parser_tagtoname.set_defaults(func=cli.run_tagtoname)
+
+# nametotag
+parser_nametotag = subparsers.add_parser(
+    "nametotag", description="Copy cell barcode sequences from read name to read tag"
+)
+parser_nametotag.add_argument(
+    "-b",
+    "--bam",
+    help="Input SAM/BAM file, '-' reads from stdin",
+    required=True,
+    type=str,
+)
+parser_nametotag.add_argument(
+    "-o",
+    "--output",
+    help="Output SAM/BAM file, '-' outputs to stdout (default '-')",
+    default="-",
+    type=str,
+)
+parser_nametotag.add_argument(
+    "-O",
+    "--outputformat",
+    help="Output format. One of 't' (SAM), 'b' (BAM),"
+    " or 'u' (uncompressed BAM) ('t' default)",
+    default="t",
+)
+parser_nametotag.add_argument(
+    "--barcode_regex",
+    help="""
+    Regular expression used to extract cell barcode from read name.
+    Default ("[^:]*") matches all characters up to the first colon.
+    """,
+    required=False,
+    type=str,
+    default="[^:]*",
+)
+parser_nametotag.add_argument(
+    "--tag",
+    help="Read tag to copy to.",
+    required=False,
+    default="CB",
+    type=str,
+)
+parser_nametotag.set_defaults(func=cli.run_nametotag)
+
+
+def main():
+    if len(sys.argv[1:]) == 0:
+        parser.print_help()
+        parser.exit()
+    else:
+        options = parser.parse_args()
+        sys.setrecursionlimit(200000)
+        options.func(options)
+
+if __name__ == "__main__":
+    main()